refactor: api types moved to shared
