+++ src/pages/api/forgotPassword/sendCode/index.ts
@@ -2,22 +2,21 @@ import { Amplify, Auth } from 'aws-amplify';
 import { amplifyConfig } from 'config/cognito';
 import { NextApiResponse } from 'next';
 import { BaseResponse } from 'services/httpClient/types';
+import { ResendCodeApiRequest } from 'shared/types/api/forgotPassword/resendCode';
 import { Logger } from 'utils';
-
-import { SendCodeApiRequest } from './types';
 
 Amplify.configure(amplifyConfig);
 
-export default async function handler<T>(
-  req: SendCodeApiRequest,
-  res: NextApiResponse<BaseResponse<T>>
+export default async function handler(
+  req: ResendCodeApiRequest,
+  res: NextApiResponse<BaseResponse<void>>
 ) {
   if (req.method === 'POST') {
     const { email } = req.body;
 
     try {
       console.log({ email });
-      const response = (await Auth.forgotPassword(email)) as T;
+      const response = await Auth.forgotPassword(email);
 
       return res.status(200).json({ data: response, ok: true, status: 200 });
     } catch (error) {
