+++ src/pages/api/forgotPassword/submit/index.ts
@@ -2,21 +2,20 @@ import { Amplify, Auth } from 'aws-amplify';
 import { amplifyConfig } from 'config/cognito';
 import { NextApiResponse } from 'next';
 import { BaseResponse } from 'services/httpClient/types';
+import { SubmitNewPasswordApiRequest } from 'shared/types/api/forgotPassword/submit';
 import { Logger } from 'utils';
-
-import { SubmitNewPasswordApiRequest } from './types';
 
 Amplify.configure(amplifyConfig);
 
-export default async function handler<T>(
+export default async function handler(
   req: SubmitNewPasswordApiRequest,
-  res: NextApiResponse<BaseResponse<T>>
+  res: NextApiResponse<BaseResponse<string>>
 ) {
   if (req.method === 'POST') {
     const { username, code, newPassword } = req.body;
 
     try {
-      const response = (await Auth.forgotPasswordSubmit(username, code, newPassword)) as T;
+      const response = await Auth.forgotPasswordSubmit(username, code, newPassword);
 
       return res.status(200).json({ data: response, ok: true, status: 200 });
     } catch (error) {
