+++ src/pages/api/resendCode/index.ts
@@ -2,21 +2,20 @@ import { Amplify, Auth } from 'aws-amplify';
 import { amplifyConfig } from 'config/cognito';
 import { NextApiResponse } from 'next';
 import { BaseResponse } from 'services/httpClient/types';
+import { ResendCodeApiRequest } from 'shared/types/api/signUp/resendCode';
 import { Logger } from 'utils';
-
-import { ResendCodeApiRequest } from './types';
 
 Amplify.configure(amplifyConfig);
 
-export default async function handler<T>(
+export default async function handler(
   req: ResendCodeApiRequest,
-  res: NextApiResponse<BaseResponse<T>>
+  res: NextApiResponse<BaseResponse<void>>
 ) {
   if (req.method === 'POST') {
     const { email } = req.body;
 
     try {
-      const response = (await Auth.resendSignUp(email)) as T;
+      const response = await Auth.resendSignUp(email);
 
       return res.status(200).json({ data: response, ok: true, status: 200 });
     } catch (error) {
