+++ src/shared/types/api/verifyEmail.ts
